Update and rename TyraMcCoyHW3.cpp to HeatTransferRate.cpp
Given Problem:
Write a C++ program that accepts four inputs: 
1) surface type as given below
2) surface dimensions
3) surface temperature (Ts)- Fahrenheit (F) or Celsius (C)
4) ambient air temperature (Ta)- Fahrenheit (F) or Celsius (C)

The program then calculates and displays the heat transfer rate through air given by q = hA(Ts − Ta), 
where q is the heat transfer rate (watts or Joules/sec), h is the convective heat transfer coefficient
(BTU/hr-ft(F) or watts/m^2(C)), and A is the surface area (ft^2 or m^2).

The three surface types are:
a. Rectangular area
b. Elliptical area
c. Other shapes

Develop this program that prompts the user for the surface type and appropriate dimensions. e.g.
The program may prompt the user to enter ’a’, ’b’ or ’c’ for each area type. If the user enters
’a’, the program should ask the user to enter the surface length (l) & width (w), and the program
calculates the area : l × w. If the user enters ’b’, the program asks the user to enter the major
and minor axis of the ellipse and calculates the surface area of the ellipse. If the user selects ’c’,
the program should ask the user to enter the surface area directly. The heat transfer rate should
then be calculated and displayed, using the convective heat transfer coefficient of 2.8 watts/m^2 Celsius,
which should be defined as the symbolic constant AIRCONV.